[refactor] 채팅방 목록 조회 url 변경
