+++ src/main/java/org/example/chat/service/ChatService.java
@@ -1,17 +1,13 @@
 package org.example.chat.service;
 
-import org.example.chat.dto.response.CreateChatRoomResponse;
 import org.example.chat.dto.request.ChatMessageRequest;
-import org.example.chat.dto.response.GetChatRoomsResponse;
+import org.example.chat.dto.response.CreateChatRoomResponse;
 import org.example.chat.repository.entity.ChatRoomEntity;
 import org.example.users.repository.entity.UserEntity;
-
-import java.util.List;
 
 public interface ChatService {
     void saveMessage(Long roomId, ChatMessageRequest request);
     CreateChatRoomResponse createChatRoom(Long productId, String chatRoomName);
-    List<GetChatRoomsResponse> getChatRooms();
-    List<String> joinChatRoom(Long roomId);
+    void joinChatRoom(Long roomId);
     void addParticipantToRoom(ChatRoomEntity chatRoom, UserEntity user);
 }
